test(InfiniteQueryBehavior): move new tests to InfiniteQueryBehavior
